feat(audio): configurable capture/playback rate on NewPortAudioSession (#1504)
Add functional-options API (WithCaptureRate/WithPlaybackRate) to
NewPortAudioSession so callers can override the default 16 kHz capture /
24 kHz playback rates. Buffer sizes are derived from the rate using named
constants (captureWindowDivisor=10 → 100 ms, playbackWindowMs=40 → 40 ms).
Defaults reproduce the pre-refactor 1600/960 frame counts exactly.

Split portaudio_session.go (testable session/option/source/sink layer) from
portaudio_io_interactive.go (hardware-bound PortAudio I/O core, excluded from
coverage threshold). The portaudioIO struct gains captureRate/playbackRate
instance fields; Start() and portaudioSource.pump() use them instead of
package constants. CaptureSampleRate/PlaybackSampleRate are retained as
documented defaults. voice.NewAudioIO() is unchanged (no args to the call).

Tests added (all hardware-free): default preservation (1600/960 locked in),
WithCaptureRate(24000)→2400 frames, WithPlaybackRate(48000)→1920 frames,
source MediaFrame.Format.SampleRate reflects the configured capture rate,
plus sink Write/Flush/Kind/Close and source Kind/Close accessors.
portaudio_session.go coverage: ~92%.
